样式格式化

+++ IFramework.Core/IO/DirectoryUtils.cs
@@ -12,9 +12,7 @@ namespace IFramework.Core
         /// </summary>
         public static void Create(string path)
         {
-            if (!Exists(path)) {
-                Directory.CreateDirectory(path);
-            }
+            if (!Exists(path)) { Directory.CreateDirectory(path); }
         }
 
         /// <summary>
@@ -22,9 +20,7 @@ namespace IFramework.Core
         /// </summary>
         public static void Remove(string path)
         {
-            if (Directory.Exists(path)) {
-                Directory.Delete(path, true);
-            }
+            if (Directory.Exists(path)) { Directory.Delete(path, true); }
         }
 
         /// <summary>
@@ -32,9 +28,7 @@ namespace IFramework.Core
         /// </summary>
         public static void Clear(string path)
         {
-            if (Directory.Exists(path)) {
-                Directory.Delete(path, true);
-            }
+            if (Directory.Exists(path)) { Directory.Delete(path, true); }
             Directory.CreateDirectory(path);
         }
 
@@ -63,18 +57,14 @@ namespace IFramework.Core
         public static void Copy(string sourcePath, string destPath, bool recursion = true)
         {
             // 如果数据源路径不存在，则抛出异常
-            if (!Directory.Exists(sourcePath)) {
-                throw new DirectoryNotFoundException(sourcePath);
-            }
+            if (!Directory.Exists(sourcePath)) { throw new DirectoryNotFoundException(sourcePath); }
             DirectoryInfo directory = new DirectoryInfo(sourcePath);
 
             //获取目录下所有文件、文件夹
             DirectoryInfo[] directories = directory.GetDirectories();
 
             // 如果目标文件夹不存在，则创建
-            if (!Directory.Exists(destPath)) {
-                Directory.CreateDirectory(destPath);
-            }
+            if (!Directory.Exists(destPath)) { Directory.CreateDirectory(destPath); }
 
             // 获取当前文件夹下所有文件
             FileInfo[] files = directory.GetFiles();
@@ -109,10 +99,7 @@ namespace IFramework.Core
         {
             List<string> fileList;
 
-            if (!recursion) {
-                fileList = Directory.GetFiles(folderPath).ToList();
-            }
-            else {
+            if (!recursion) { fileList = Directory.GetFiles(folderPath).ToList(); } else {
                 fileList = Directory.GetFiles(folderPath).ToList();
 
                 //找出所有子文件夹
@@ -143,9 +130,7 @@ namespace IFramework.Core
                 return null;
             }
 
-            if (directory.Parent != null && directory.Attributes.ToString().IndexOf("System", StringComparison.Ordinal) > -1) {
-                return null;
-            }
+            if (directory.Parent != null && directory.Attributes.ToString().IndexOf("System", StringComparison.Ordinal) > -1) { return null; }
             FileInfo[] fileInfos = directory.GetFiles(fileName);
             fileList.AddRange(fileInfos?.Select(file => file.FullName));
